// src/shared/types/operador.ts
export interface UsuarioViewModel {
    op_codigo: number;
    op_nombre: string;
    op_documento: string;
    op_rol: string;
    op_sucursal: string;
    op_autorizar: number;
    op_ver_utilidad: number;
    op_ver_proveedor: number;
    op_aplicar_descuento: number;
    op_movimiento: number;
}